improv: delay the logging timeout when calling starting it multiple times
The 2 minutes timeout is a way to prevent performance overhead if the
frontend doesn't stop the logging (ex: the user closes the tab).  To
have logs continuously displayed in the frontend, we should be able to
start again the logs when it timeouts.  But right now, the frontend
can't predict when it'll timeout, because the timeout may happens before
the 2 minutes if logs were already started.

Postponing the logging timeout will allow the front to safely restart
logging every 2 minutes when needed.